// lib/api/generated/models/api-v1-room-viewings-get200-response-data-inner.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Rental Search API V1
 * API for searching rental rooms on map with PostGIS support
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface ApiV1RoomViewingsGet200ResponseDataInner
 */
export interface ApiV1RoomViewingsGet200ResponseDataInner {
    /**
     * 
     * @type {number}
     * @memberof ApiV1RoomViewingsGet200ResponseDataInner
     */
    'id'?: number;
    /**
     * 
     * @type {number}
     * @memberof ApiV1RoomViewingsGet200ResponseDataInner
     */
    'room_id'?: number;
    /**
     * 
     * @type {string}
     * @memberof ApiV1RoomViewingsGet200ResponseDataInner
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof ApiV1RoomViewingsGet200ResponseDataInner
     */
    'email'?: string;
    /**
     * 
     * @type {string}
     * @memberof ApiV1RoomViewingsGet200ResponseDataInner
     */
    'phone'?: string;
    /**
     * 
     * @type {string}
     * @memberof ApiV1RoomViewingsGet200ResponseDataInner
     */
    'preferred_date'?: string;
    /**
     * 
     * @type {string}
     * @memberof ApiV1RoomViewingsGet200ResponseDataInner
     */
    'message'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ApiV1RoomViewingsGet200ResponseDataInner
     */
    'status'?: string;
    /**
     * 
     * @type {string}
     * @memberof ApiV1RoomViewingsGet200ResponseDataInner
     */
    'created_at'?: string;
    /**
     * 
     * @type {string}
     * @memberof ApiV1RoomViewingsGet200ResponseDataInner
     */
    'updated_at'?: string;
}